refactor: move historial and currentCrews to profile file
